Day 3 complete
Added a tool to generate files from boilerplates.

+++ 2016/c/sxc_vector.h
@@ -48,7 +48,7 @@ enum SXC_Vector_scalars {
 
 #define sxc_vector_find(v, tar, cmp, found)				\
 	do {								\
-		for (int i = 0; i < (v).size; ++i) {			\
+		for (size_t i = 0; i < (v).size; ++i) {			\
 			if (cmp( &((v).vec[i]), tar)) {			\
 				found = &((v).vec[i]);			\
 				break;					\
